rename private mod towkt -> to_wkt
