+++ src/lib.rs
@@ -74,8 +74,8 @@ use crate::types::MultiPolygon;
 use crate::types::Point;
 use crate::types::Polygon;
 
+mod to_wkt;
 mod tokenizer;
-mod towkt;
 
 /// `WKT` primitive types and collections
 pub mod types;
@@ -85,7 +85,7 @@ extern crate geo_types;
 
 extern crate thiserror;
 
-pub use crate::towkt::ToWkt;
+pub use crate::to_wkt::ToWkt;
 
 #[cfg(feature = "geo-types")]
 #[deprecated(note = "renamed module to `wkt::geo_types_from_wkt`")]
